QCMPLUS-29 - Mise à jour des opérations CRUD avec Axios et gestion des erreurs and Ajout d'un gestionnaire d'exceptions utilisateur côté backend

+++ src/main/java/com/qcmplus/qcmplus/controller/AdminController.java
@@ -1,20 +1,19 @@
 package com.qcmplus.qcmplus.controller;
 
+import com.qcmplus.qcmplus.exception.AdminEmailAlreadyInUseException;
+import com.qcmplus.qcmplus.exception.AdminNotFoundException;
 import com.qcmplus.qcmplus.model.Admin;
 import com.qcmplus.qcmplus.service.AdminService;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Optional;
 
 @RestController
+@RequestMapping("/admin") // New request mapping at the controller level.
 public class AdminController {
 
     private final AdminService adminService;
@@ -24,40 +23,46 @@ public class AdminController {
         this.adminService = adminService;
     }
 
-    @GetMapping("/admin")
-    public List<Admin> getAllAdmins() {
-        return adminService.getAllAdmins();
+    @GetMapping
+    public ResponseEntity<List<Admin>> getAllAdmins() {
+        return new ResponseEntity<>(adminService.getAllAdmins(), HttpStatus.OK);
     }
 
-    @GetMapping("admin/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<Admin> getAdminById(@PathVariable Long id) {
         Optional<Admin> admin = adminService.getAdminById(id);
         return admin.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PostMapping("/admin")
+    @PostMapping
     public ResponseEntity<Admin> createAdmin(@RequestBody Admin admin) {
         try {
             Admin savedAdmin = adminService.saveAdmin(admin);
             return new ResponseEntity<>(savedAdmin, HttpStatus.CREATED);
-        } catch (Exception e) {
+        } catch (AdminEmailAlreadyInUseException e) {
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
     }
 
-    @PostMapping("/admin/{id}")
-    public ResponseEntity<Admin> updateAdmin(@PathVariable Long id, @RequestBody Admin adminDetails) {
+    @PostMapping("/{id}")
+    public ResponseEntity<?> updateAdmin(@PathVariable Long id, @RequestBody Admin adminDetails) {
         try {
             Admin updatedAdmin = adminService.updateAdmin(id, adminDetails);
             return ResponseEntity.ok(updatedAdmin);
+        } catch (AdminEmailAlreadyInUseException e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred.");
         }
     }
 
-    @DeleteMapping("/admin/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteAdmin(@PathVariable Long id) {
-        adminService.deleteAdmin(id);
-        return new ResponseEntity<>("Admin with id " + id + " deleted successfully", HttpStatus.OK);
+        if(adminService.existsById(id)) {
+            adminService.deleteAdmin(id);
+            return new ResponseEntity<>("Admin with ID " + id + " deleted successfully", HttpStatus.OK);
+        } else {
+            throw new AdminNotFoundException(Math.toIntExact(id));
+        }
     }
 }
